use std::size_t where appropriate
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@6492 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/util/dstring.h
@@ -50,7 +50,7 @@ public:
     return no==0; // string 0 is exactly the empty string
   }
   
-  inline char operator[](unsigned i) const
+  inline char operator[](size_t i) const
   {
     return as_string()[i];
   }
